LayoutTests/imported/w3c/web-platform-tests/html/dom/self-origin.any-expected.txt


FAIL self.origin should be correct assert_equals: expected (string) "http://localhost:8800" but got (undefined) undefined
